requirements.txt
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
ipaddress==1.0.23
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
netns==1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.3.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyroute2==0.7.9
scapy==2.5.0
setuptools==65.5.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
wheel==0.38.4
PyYAML~=6.0.1